feat : using Short for version instead of Long

+++ graphql/boot-graphql-webmvc/src/main/java/com/example/graphql/entities/AuthorEntity.java
@@ -45,7 +45,7 @@ public class AuthorEntity implements Serializable {
 
     private LocalDateTime registeredAt;
 
-    @Version private Long version;
+    @Version private Short version;
 
     @OneToMany(mappedBy = "authorEntity", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<PostEntity> postEntities = new ArrayList<>();
@@ -85,7 +85,7 @@ public class AuthorEntity implements Serializable {
         return this;
     }
 
-    public AuthorEntity setVersion(Long version) {
+    public AuthorEntity setVersion(Short version) {
         this.version = version;
         return this;
     }
