+++ jpa/multitenancy/multidatasource-multitenancy/src/main/java/com/example/multitenancy/primary/entities/PrimaryCustomer.java
@@ -34,7 +34,7 @@ public class PrimaryCustomer {
     @Setter(AccessLevel.NONE)
     @Version
     @Column(name = "version")
-    private Long version;
+    private Short version;
 
     @Column(nullable = false)
     @TenantId
