+++ r2dbc/boot-jooq-r2dbc-sample/src/main/java/com/example/jooq/r2dbc/entities/Post.java
@@ -53,5 +53,5 @@ public class Post {
 
     @Column("version")
     @Version
-    private Long version;
+    private Short version;
 }
