+++ r2dbc/boot-jooq-r2dbc-sample/src/main/resources/db/migration/postgresql/V1__01_init.sql
@@ -9,7 +9,7 @@ CREATE TABLE posts
     created_at timestamptz DEFAULT NOW(),
     created_by text,
     updated_at timestamptz,
-    version BIGINT DEFAULT 0,
+    version int DEFAULT 0,
     PRIMARY KEY   (ID)
 );
